refactor: split LSP wiring from plugin spec and tidy init.lua
Three related structural cleanups bundled:

- lua/config/lsp.lua (new): extract all LSP wiring — vim.lsp.config,
  vim.lsp.enable, and the LspAttach autocmd with its keymaps — out
  of the nvim-lspconfig plugin spec. The lsp.lua plugin file is now
  a three-line shim that just event-triggers and requires the new
  config module. Keeps lazy loading semantics unchanged while
  separating "which plugin manager owns lspconfig" from "how this
  config wires up LSP".

- colorscheme call moves from init.lua to pack.lua, right after
  require("catppuccin").setup(). Cohesion win — setup and apply are
  now next to each other, and init.lua is purely a require sequence.

- jumpoptions gains "view": preserves viewport (scroll position and
  cursor row) when navigating the jumplist with <C-o>/<C-i>. Nvim
  0.11+ feature, pure QoL.
