+++ init.lua
@@ -2,5 +2,3 @@ require("config.options")
 require("config.pack")
 require("config.lazy")
 require("config.keymaps")
-
-vim.cmd([[colorscheme catppuccin]])
